refactor(memory): move the remaining inert DTOs into tinycortex-api
