+++ api/Cargo.toml
@@ -10,9 +10,27 @@ description = "Stable public contracts for the TinyCortex memory system"
 # hosts compile against, so it must stay free of native, async-runtime, and
 # storage dependencies. Anything heavier belongs in the `tinycortex` engine
 # crate, never here.
+#
+# The full set is intentionally small and pure-Rust. Beyond the
+# serde/error/async-trait baseline it carries exactly three additions, each
+# pulled in by a value type that has to keep behaving identically after the
+# move out of the engine crate:
+#
+# - `chrono`  — timestamps on chunk/tree nodes; the `serde` feature backs
+#               `chunks::Metadata`'s `chrono::serde::ts_milliseconds`.
+# - `sha2`    — the deterministic `chunks::chunk_id`.
+# - `uuid`    — `tool_memory::ToolMemoryRule::generate_id` (v4 bytes, nibble
+#               encoded). Only the `v4` feature is needed here; the engine
+#               crate additionally enables `serde`.
+#
+# Nothing here may pull in `rusqlite`, `git2`, `reqwest`, `regex`, or an async
+# runtime — guard with `cargo tree -i <crate> -p tinycortex-api`.
 [dependencies]
 anyhow = "1"
 async-trait = "0.1"
+chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+sha2 = "0.10"
 thiserror = "2"
+uuid = { version = "1", features = ["v4"] }
